tmaurer3/move_code_from_h_to_cpp
